renaming card to item
